requirements.txt
Django==4.0.2
django-allauth==0.51.0
Pillow==9.0.1